Move test registry into `sigstoretest` package
