+++ .golangci.yml
@@ -173,7 +173,7 @@ linters-settings:
           - '!**/lib/srv/mock.go'
           - '!**/lib/srv/db/redis/test.go'
           - '!**/lib/tbot/workloadidentity/workloadattest/podman/test_server.go'
-          - '!**/lib/tbot/workloadidentity/workloadattest/sigstore/testutils.go'
+          - '!**/lib/tbot/workloadidentity/workloadattest/sigstore/sigstoretest/sigstoretest.go'
           - '!**/lib/teleterm/gatewaytest/**'
           - '!**/lib/utils/testhelpers.go'
           - '!**/lib/utils/testutils/**'
@@ -260,7 +260,7 @@ linters-settings:
           - '!**/lib/srv/mock.go'
           - '!**/lib/srv/db/redis/test.go'
           - '!**/lib/tbot/workloadidentity/workloadattest/podman/test_server.go'
-          - '!**/lib/tbot/workloadidentity/workloadattest/sigstore/testutils.go'
+          - '!**/lib/tbot/workloadidentity/workloadattest/sigstore/sigstoretest/sigstoretest.go'
           - '!**/lib/teleterm/gatewaytest/**'
           - '!**/lib/utils/cli.go'
           - '!**/lib/utils/testhelpers.go'
