+++ lib/tbot/workloadidentity/workloadattest/sigstore/sigstoretest/testdata/v2/simple-signing/blobs/sha256~608c32cf47678ef8e63578597fbb982ddf3b84cb491359e7c6e824b3925a8ce2
