+++ lib/tbot/workloadidentity/workloadattest/sigstore/sigstoretest/testdata/v2/simple-signing/manifests/sha256-21c76c650023cac8d753af4cb591e6f7450c6e2b499b5751d4a21e26e2fc5012.sig
